Forward-port of PR odoo/odoo#86643.

Add a unittest for multi_step route on Ship Later warehouse to
ensure that the fix https://github.com/odoo/odoo/commit/5959d77f3448eb4cafabeb744a94b21471353e02 is not unintentionally reverted.


#### speed

As for the v14 PR the speed depends on a number of things. First in a customer DB with no products tracked,
realtime inventory valuation and no Bills of materials. Speedup when closing different sessions:

- 149 orders, 346 lines, 288 products: 410s -> 380s
- 381 orders, 1048 lines, 687 products: 766s -> 301s
- 572 orders, 1395 lines, 830 products: 536s -> 367s

Then, in a testing database with no products tracked, no anglo-saxon accounting, real-time inventory valuation,
phantom boms for each product with 5-30 components:

- 70 orders, 350 lines, 350 products: 32 min -> 13 min.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr