@pimodoo 2022-04-05
Did you have the opportunity to test the performances gain in this version? As the ORM, as changed a lot, in his version, I wonder if creating in batch is really faster?