@Aurelienvd 2022-04-06
@pimodoo I'll benchmark a bit on a v15 customer DB to check the possible perf gain.