@Aurelienvd 2022-04-07
@pimodoo Did some benchmarks. The speed-up is not quite the same as the one in the v14 PR, although
it may be linked to some config settings that I did not replicate. Still, it seems to give a 2-3x speed-up on average.